+++ core/.gitignore
@@ -9,4 +9,7 @@ db.sqlite3
 
 prod.env
 .env.dev
+.env.local
 *.env
+
+
