+++ core/core/urls.py
@@ -3,5 +3,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("api/", include("food.urls"))
+    path("api/", include("food.urls")),
+    path("api/orders/", include("orders.urls")),
 ]
