+++ core/docker-compose.yml
@@ -12,3 +12,16 @@ services:
       - 8000:8000
     env_file:
       - ./.env
+    depends_on:
+      - db
+
+  db:
+    image: postgres:13.0-alpine
+    volumes:
+      - postgres_data:/var/lib/postgresql/data/
+    env_file:
+      - ./.env
+
+volumes:
+  postgres_data:
+
