+++ core/food/urls.py
@@ -1,12 +1,12 @@
 from django.urls import path
 from rest_framework.routers import DefaultRouter
 
-from .views import NomenclatureViewSet, MenuViewSet, LDAPLogin, OrderViewSet
+from .views import NomenclatureViewSet, MenuViewSet, LDAPLogin
+from orders.views import OrderViewSet
 
 router = DefaultRouter()
 router.register(r'nomenclature', NomenclatureViewSet, basename='nomenclature')
 router.register(r'menu', MenuViewSet, basename='menu')
-router.register(r'orders', OrderViewSet, basename='orders')
 
 urlpatterns = router.urls
