+++ core/orders/admin.py
@@ -1,3 +1,15 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import OrderContent, Order
+
+
+class OrderContentInLine(admin.TabularInline):
+    model = OrderContent
+
+
+@admin.register(Order)
+class OrderAdmin(admin.ModelAdmin):
+    list_display = ('id', 'created_by', 'created_at', )
+    inlines = [
+        OrderContentInLine
+    ]
