+++ core/orders/urls.py
@@ -1,0 +1,8 @@
+from rest_framework.routers import DefaultRouter
+
+from .views import OrderViewSet
+
+router = DefaultRouter()
+router.register(r'', OrderViewSet, basename='orders')
+
+urlpatterns = router.urls
