+++ core/requirements.txt
@@ -2,6 +2,7 @@ asgiref==3.7.2
 Django==5.0.1
 django-auth-ldap==4.6.0
 django-cors-headers==4.3.1
+django-filter==23.5
 djangorestframework==3.14.0
 psycopg2-binary==2.9.9
 pyasn1==0.5.1
